+++ transistor.cpp
@@ -7,9 +7,9 @@ void Transistor::Transistor(string alias,
 						string gate,
 						string bulk,
 						string type, //PMOS,NMOS,LVT,HVT
-						float diff_width,
+						double diff_width,
 						int fingers,
-						float gate_lenght,
+						double gate_lenght,
 						int stack){
 
 }
@@ -38,7 +38,7 @@ void set_type(string type){
 
 }
 
-void set_diff_width(float width){
+void set_diff_width(double width){
 
 }
 
@@ -46,7 +46,7 @@ void set_fingers(int fingers){
 
 }
 
-void set_gate_lenght(float gate_lenght){
+void set_gate_lenght(double gate_lenght){
 
 }
 
@@ -79,7 +79,7 @@ string get_type( ){
 
 }
 
-float get_diff_width( ){
+double get_diff_width( ){
 
 }
 
@@ -87,7 +87,7 @@ int get_fingers( ){
 
 }
 
-float get_gate_lenght( ){
+double get_gate_lenght( ){
 
 }
